add setting to limit profile editing by non-admins
Test Plan:
1. Validate that a non-admin user can modify their name, profile links,
    bio and communication channels
2. Disable the respective settings in the Account settings tab
3. Validate that the non-admin user can no longer edit the listed items
4. Validate that the API does not allow these changes either

flag=none

refs PFS-20887

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/304025
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Vaughn Hepworth <vhepworth@instructure.com>
QA-Review: Douglas Rist <douglas.rist@instructure.com>

+++ app/controllers/profile_controller.rb
@@ -423,13 +423,13 @@ class ProfileController < ApplicationController
 
     short_name = params[:user] && params[:user][:short_name]
     @user.short_name = short_name if short_name && @user.user_can_edit_name?
-    if params[:user_profile]
+    if params[:user_profile] && @user.user_can_edit_profile?
       user_profile_params = params[:user_profile].permit(:title, :bio)
       user_profile_params.delete(:title) unless @user.user_can_edit_name?
       @profile.attributes = user_profile_params
     end
 
-    if params[:link_urls] && params[:link_titles]
+    if params[:link_urls] && params[:link_titles] && @user.user_can_edit_profile?
       @profile.links = []
       params[:link_urls].zip(params[:link_titles])
                         .reject { |url, title| url.blank? && title.blank? }
